4o mini - copia/requirements.txt
fastapi
uvicorn[standard]
openai
python-dotenv
pydantic
websockets
edge-tts
sounddevice
numpy
pydub
chromadb
tiktoken
